+++ pysys-examples/internal/testcases/PySys_internal_102/run.py
@@ -1,4 +1,4 @@
-import os, sys, math, shutil, glob, io
+import os, sys, math, shutil, glob, io, zipfile
 from pysys.constants import *
 from pysys.basetest import BaseTest
 
@@ -10,7 +10,8 @@ class PySysTest(BaseTest):
 
 	def execute(self):
 		self.copy(self.input, self.output+'/test')
-		runPySys(self, 'pysys', ['run', '-o', self.output+'/pysys-output', '--purge'], workingDir='test')
+		runPySys(self, 'pysys', ['run', '-o', self.output+'/pysys-output', '--purge'], workingDir='test', 
+			onError=lambda process: [self.logFileContents(process.stdout), self.logFileContents(process.stderr)] )
 		# run a recond time to prove earlier files aren't kept
 		runPySys(self, 'pysys', ['run', '-o', self.output+'/pysys-output', '--purge', '--cycle', '2'], workingDir='test')
 		self.logFileContents('pysys.out', maxLines=0)
@@ -22,8 +23,28 @@ class PySysTest(BaseTest):
 	def validate(self):
 		self.logFileContents('collected_files.txt')
 		self.assertDiff('collected_files.txt', 'ref_collected_files.txt')
-		self.assertGrep('pysys.out', expr=r'Collected test output to directory: .*mydir-pysys-output$')
-		self.assertGrep('pysys.out', expr=r'Collected test output to directory: .*mydir2[/\\]\d\d\d\d-\d\d-\d\d_\d\d.\d\d.\d\d')
+		self.assertGrep('pysys.out', expr=r'Collected .+ test output files to directory: .*mydir-pysys-output$')
+		self.assertGrep('pysys.out', expr=r'Collected .+ test output files to directory: .*mydir2[/\\]\d\d\d\d-\d\d-\d\d_\d\d.\d\d.\d\d')
 		mydirs = glob.glob(self.output+'/mydir2/*')[0]
 		self.assertPathExists(mydirs+'/1.myfile')
 		self.assertPathExists(mydirs+'/2.myfile')
+		
+		self.assertPathExists('pysys-output/mywriter_defaultpattern/NestedTest.cycle001.a-foo.1.myext') # extension at the end after unique id
+		
+		# -b excluded by regex filter
+		self.assertPathExists('pysys-output/mywriter_defaultpattern/NestedTest.cycle001.foo-b.1')
+		self.assertThat('actual == []', actual__eval='import_module("glob").glob(self.output+"/pysys-output/mywriter-pysys-output/mydir/*foo-b*")')
+		self.assertPathExists('pysys-output/mywriter-pysys-output/mydir/NestedTest.cycle001.a-foo.1.myext')
+		
+		with zipfile.ZipFile(self.output+'/pysys-output/mywriter-pysys-output/myArchive.zip') as zf:
+			self.assertThat('badZipFiles is None', badZipFiles=zf.testzip())
+			members = sorted(zf.namelist())
+		self.assertThat('members == expected', members=sorted(members), expected=sorted([
+			'mydir/NestedTest.cycle001.a-foo.1.myext', 
+			'mydir/NestedTest.cycle001.a-foo.2.myext', 
+			'mydir/NestedTest.cycle001.foo.1', 
+			'mydir/NestedTest.cycle001.foo.2', 
+			'mydir/NestedTest.cycle002.a-foo.1.myext', 
+			'mydir/NestedTest.cycle002.a-foo.2.myext', 
+			'mydir/NestedTest.cycle002.foo.1', 
+			'mydir/NestedTest.cycle002.foo.2']))